requirements.txt
fastapi
uvicorn[standard]
streamlit

langchain
langchain-core
langchain-ollama
sentence-transformers
openai

redis

httpx

PyJWT

python-dotenv
pydantic